requirements.txt
# UrbanSAR - Multi-Sensor Fusion for Building Height Estimation
# Core ML
torch>=2.0.0
torchvision>=0.15.0

# Geospatial
rasterio>=1.3.0
rioxarray>=0.15.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
fiona>=1.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
Pillow>=10.0.0

# Dashboard
streamlit>=1.28.0
folium>=0.15.0
streamlit-folium>=0.15.0
plotly>=5.18.0

# Utilities
kaggle>=1.5.0
tqdm>=4.65.0
matplotlib>=3.7.0
